+++ hastexo/common.py
@@ -174,6 +174,8 @@ DEFAULT_SETTINGS = {
     },
     "providers": {},
     "guacamole_js_version": '1.4.0',
+    "lab_usage_limit": None,
+    "lab_usage_limit_breach_policy": None
 }